refactor: separate cloning and init list parsing
